Rename store to currentCard
