+++ src/stores/currentCard.ts
